+++ CHANGELOG.md
@@ -12,6 +12,7 @@ This changelog's format is based on [keep a changelog 1.0.0](https://keepachange
 - Added `--parse-permissive` parsing mode for traces
 - `monkey` now performs IPASIR lookups within the current process' own symbols if the specified DSO name is `preloaded`
 - Added support for dynamically linking a fixed IPASIR solver to `monkey` via the `IM_IPASIR_LIB` build option
+- Added `--crash-on-failure` failure mode for `monkey replay`
 
 ### Changed